backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
# psycopg2-binary==2.9.9  # Only needed for PostgreSQL - not compatible with Python 3.14
# Use psycopg (psycopg3) for Python 3.14+ with PostgreSQL: pip install psycopg[binary]
psycopg[binary]>=3.0.0  # PostgreSQL adapter for Python 3.14+
Pillow>=10.2.0
python-dateutil==2.8.2
reportlab==4.0.7
python-docx==1.1.0
PyPDF2==3.0.1
django-extensions==3.2.3
cryptography>=41.0.0
django-auditlog==3.0.0
djangorestframework-simplejwt==5.3.0
django-storages==1.14.2
boto3==1.34.0
waitress>=2.1.2
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0


